requirements.txt
pandas
duckdb
langchain-chroma
langchain-google-genai
python-dotenv
openpyxl
tqdm
